zebra: Store received NHG from upper level protocol
Upper level protocols are sending down a list of nexthops
that a temporary nhg was created to hold it.  After
route resolution is done, this NHG is thrown away.
Let's keep this received NHG as a pointer off the RE.

Follow the `installedNexthopGroupId` pattern for display
purposes at this point in time.

Signed-off-by: Donald Sharp <sharpd@nvidia.com>