vtysh: Gather up failure return codes
Currently vtysh when working with mgmtd is not
properly gathering up the return codes to pass
them back to the person initiating the command.
This can be seen by issuing something like this:

failure.conf:
int lo
  ip address 0.0.0.5/32

vtysh -f failure.conf

With this change vtysh now notices that this fails
and let's the operator know that it did via
the return code.

Signed-off-by: Donald Sharp <sharpd@nvidia.com>