*: make regex_t opaque for API isolation
Directly embedding `regex_t` in `struct vty` creates a problem for
anyone trying to use a vty in external code referring to `lib/vty.h` -
without knowing which regex library FRR was built against, the size of
`regex_t` can't be known, and so the struct layout is unknown.

=> encapsulate `regex_t` in a dynamically allocated struct everywhere
it's used, so it's always just a pointer.

Signed-off-by: David Lamparter <equinox@opensourcerouting.org>