bgpd: Change name for bgp_evpn_route_add_l3_ecomm_ok
The conditions in this API are applicable for both L3VNI related extended
communities as well as L3VNI related NLRI embedded attributes such as
MPLS Label 2. Renaming to bgp_evpn_route_add_l3_attrs_ok so the API can
be used to determine if we need to attach the second label to the mac-ip
route.

Ticket: #4413589

Signed-off-by: Mitesh Kanjariya <mkanjariya@nvidia.com>