bgp_bmp: do not send end of rib by default
Currently, BMP sends an EoR for each supported policy type (Adj-RIB-In
pre-policy, Adj-RIB-In post-policy, and Loc-RIB), regardless of the
policy configuration. This may be confusing when a user only wants to
monitor Adj-RIB-In pre-policy routes but still receives an EoR for
Adj-RIB-In post-policy.

Do not send bmp EoR if the policy was not configured.

Signed-off-by: Loïc Sang <loic.sang@6wind.com>